Defer importing markdown
